+++ packages/system/src/compose.test.ts
@@ -6,6 +6,7 @@ import { color } from "./color";
 import { describe, expect, test, beforeEach } from "@jest/globals";
 import { flex } from "./flex";
 import { shadow } from "./shadow";
+import { list } from "./list";
 describe("compose function", () => {
   test("should combine styles from multiple style functions", () => {
     // Arrange
@@ -15,7 +16,8 @@ describe("compose function", () => {
       layout,
       color,
       flex,
-      shadow
+      shadow,
+      list
     );
     const props: StyledProps = {
       m: 8,
@@ -26,7 +28,9 @@ describe("compose function", () => {
       color: "red",
       flexDir: ["column", "row"],
       boxShadow: "12px 12px 2px 1px rgba(0, 0, 255, .2)",
+      listStyle: "square",
       zIndex: 9999,
+      cursor: "pointer",
     };
     // Act
     const styles = combinedFunction(props);
@@ -35,13 +39,15 @@ describe("compose function", () => {
     expect(styles.base).toContain("margin: 8px");
     expect(styles.base).toContain("font-size: 16px");
     expect(styles.base).toContain("width: 100%");
-    expect(styles.base).toContain("background-color: red");
+    expect(styles.base).toContain("background: red");
     expect(styles.base).toContain("opacity: 0.5;");
     expect(styles.base).toContain("flex-direction: column");
     expect(styles.base).toContain(
       "box-shadow: 12px 12px 2px 1px rgba(0, 0, 255, .2)"
     );
+    expect(styles.base).toContain("list-style: square;");
     expect(styles.base).toContain("z-index: 9999;");
+    expect(styles.base).toContain("cursor: pointer;");
   });
 
   test("should not include invalid keys in the resulting CSS", () => {