+++ packages/babel-plugin/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @kuma-ui/babel-plugin
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @kuma-ui/system@0.4.1
+
+## 0.4.0
+
+### Minor Changes
+
+- improve Style Props
+
+### Patch Changes
+
+- Updated dependencies
+  - @kuma-ui/system@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes