+++ packages/system/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kuma-ui/system
 
+## 0.4.1
+
+### Patch Changes
+
+- Add 'cursor' CSS support
+
+## 0.4.0
+
+### Minor Changes
+
+- improve Style Props
+
 ## 0.3.0
 
 ### Minor Changes